Create netex_utils module with 'get_value_in_keylist'
