+++ src/hellogo_fares/utils.rs
@@ -14,9 +14,7 @@
 // along with this program.  If not, see
 // <http://www.gnu.org/licenses/>.
 
-use crate::minidom_utils::TryOnlyChild;
-use crate::objects::Date;
-use crate::Result;
+use crate::{minidom_utils::TryOnlyChild, netex_utils, objects::Date, Result};
 use chrono::NaiveDate;
 use failure::{bail, format_err, Error};
 use minidom::Element;
@@ -56,32 +54,6 @@ impl FromStr for FrameType {
             _ => bail!("Failed to convert '{}' into a FrameType", s),
         }
     }
-}
-
-pub fn get_value_in_keylist<F>(element: &Element, key: &str) -> Result<F>
-where
-    F: FromStr,
-{
-    let values = element
-        .try_only_child("KeyList")?
-        .children()
-        .filter(|key_value| match key_value.try_only_child("Key") {
-            Ok(k) => k.text() == key,
-            _ => false,
-        })
-        .map(|key_value| key_value.try_only_child("Value"))
-        .collect::<Result<Vec<_>>>()?;
-    if values.len() != 1 {
-        bail!(
-            "Failed to find a unique key '{}' in '{}'",
-            key,
-            element.name()
-        )
-    }
-    values[0]
-        .text()
-        .parse()
-        .map_err(|_| format_err!("Failed to get the value out of 'KeyList' for key '{}'", key))
 }
 
 pub fn get_amount_units_factor(element: &Element) -> Result<Decimal> {
@@ -151,7 +123,8 @@ pub fn get_frame_type(frame: &Element) -> Result<FrameType> {
     let fare_structure = frame
         .try_only_child("fareStructures")?
         .try_only_child("FareStructure")?;
-    let frame_type: FrameType = get_value_in_keylist(fare_structure, "FareStructureType")?;
+    let frame_type: FrameType =
+        netex_utils::get_value_in_keylist(fare_structure, "FareStructureType")?;
     Ok(frame_type)
 }
 
@@ -210,59 +183,6 @@ mod tests {
         #[should_panic(expected = "Failed to convert \\'NotAFrameType\\' into a FrameType")]
         fn parse_invalid() {
             "NotAFrameType".parse::<FrameType>().unwrap();
-        }
-    }
-
-    mod value_in_keylist {
-        use super::super::get_value_in_keylist;
-        use minidom::Element;
-        use pretty_assertions::assert_eq;
-
-        #[test]
-        fn has_value() {
-            let xml = r#"<root>
-                    <KeyList>
-                        <KeyValue>
-                            <Key>key</Key>
-                            <Value>42</Value>
-                        </KeyValue>
-                    </KeyList>
-                </root>"#;
-            let root: Element = xml.parse().unwrap();
-            let value: u32 = get_value_in_keylist(&root, "key").unwrap();
-            assert_eq!(value, 42);
-        }
-
-        #[test]
-        #[should_panic(expected = "Failed to find a child \\'KeyList\\' in element \\'root\\'")]
-        fn no_keylist_found() {
-            let xml = r#"<root />"#;
-            let root: Element = xml.parse().unwrap();
-            get_value_in_keylist::<u32>(&root, "key").unwrap();
-        }
-
-        #[test]
-        #[should_panic(expected = "Failed to find a unique key \\'key\\' in \\'root\\'")]
-        fn no_key_found() {
-            let xml = r#"<root>
-                    <KeyList />
-                </root>"#;
-            let root: Element = xml.parse().unwrap();
-            get_value_in_keylist::<u32>(&root, "key").unwrap();
-        }
-
-        #[test]
-        #[should_panic(expected = "Failed to find a child \\'Value\\' in element \\'KeyValue\\'")]
-        fn no_value_found() {
-            let xml = r#"<root>
-                    <KeyList>
-                        <KeyValue>
-                            <Key>key</Key>
-                        </KeyValue>
-                    </KeyList>
-                </root>"#;
-            let root: Element = xml.parse().unwrap();
-            get_value_in_keylist::<u32>(&root, "key").unwrap();
         }
     }
